requirements.txt
fastapi
uvicorn
paho-mqtt
python-osc
opencv-python
